+++ TimeClock-MySQL/src/hu/gaborkolozsy/timeclock/dao/impl/PayInfoRepositoryJDBCImpl.java
@@ -2,18 +2,18 @@
  * Copyright (c) 2016, by Gábor Kolozsy. All rights reserved.
  * 
  */
-package timeclock.dao;
+package hu.gaborkolozsy.timeclock.dao.impl;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import timeclock.interfaces.PayInfoRepository;
-import timeclock.job.Job;
-import timeclock.query.PayInfo;
+import hu.gaborkolozsy.timeclock.daos.PayInfoRepository;
+import hu.gaborkolozsy.timeclock.model.Job;
+import hu.gaborkolozsy.timeclock.model.PayInfo;
 
 /**
- * This object implementing the {@link timeclock.interfaces.PayInfoRepository}
+ * This object implementing the {@link hu.gaborkolozsy.timeclock.daos.PayInfoRepository}
  * generic interface. 
  * <p>
  * It manages the relationship between the {@code PayInfo} 
@@ -24,9 +24,9 @@ import timeclock.query.PayInfo;
  * 
  * @author Kolozsy Gábor
  * @version 2.1
- * @see timeclock.job.Job
- * @see timeclock.query.PayInfo
- * @see timeclock.interfaces.PayInfoRepository
+ * @see hu.gaborkolozsy.timeclock.model.Job
+ * @see hu.gaborkolozsy.timeclock.model.PayInfo
+ * @see hu.gaborkolozsy.timeclock.daos.PayInfoRepository
  * @see java.sql.Connection
  * @see java.sql.PreparedStatement
  * @see java.sql.ResultSet
@@ -98,13 +98,13 @@ public class PayInfoRepositoryJDBCImpl implements PayInfoRepository<PayInfo,Job>
     }
     
     /**
-     * <strong>See >>></strong> {@link timeclock.interfaces.PayInfoRepository}.
+     * <strong>See >>></strong> {@link hu.gaborkolozsy.timeclock.daos.PayInfoRepository}.
      * 
      * @param e for identification the correct {@code Job}
      * 
      * @return a new {@code PayInfo} object
      * @throws SQLException 
-     * @see timeclock.interfaces.PayInfoRepository
+     * @see hu.gaborkolozsy.timeclock.daos.PayInfoRepository
      */
     @Override
     public PayInfo getInfo(Job e) throws SQLException {
@@ -125,14 +125,14 @@ public class PayInfoRepositoryJDBCImpl implements PayInfoRepository<PayInfo,Job>
     }
     
     /**
-     * <strong>See >>></strong> {@link timeclock.interfaces.PayInfoRepository}.
+     * <strong>See >>></strong> {@link hu.gaborkolozsy.timeclock.daos.PayInfoRepository}.
      * 
      * @param job for identification the correct {@code Job}
      * @param status for identification the correct {@code Job}
      * 
      * @return a new {@code PayInfo} object
      * @throws SQLException 
-     * @see timeclock.interfaces.PayInfoRepository
+     * @see hu.gaborkolozsy.timeclock.daos.PayInfoRepository
      * @deprecated <strong>not use this method</strong>
      */
     @Deprecated
@@ -146,13 +146,13 @@ public class PayInfoRepositoryJDBCImpl implements PayInfoRepository<PayInfo,Job>
     }
     
     /**
-     * <strong>See >>></strong> {@link timeclock.interfaces.PayInfoRepository}.
+     * <strong>See >>></strong> {@link hu.gaborkolozsy.timeclock.daos.PayInfoRepository}.
      * 
      * @param job for identification the correct {@code Job}
      * 
      * @return {@code hourlyPay} data member of {@code PayInfo} object
      * @throws SQLException 
-     * @see timeclock.interfaces.PayInfoRepository
+     * @see hu.gaborkolozsy.timeclock.daos.PayInfoRepository
      * @deprecated <strong>not use this method</strong>
      */
     @Deprecated
@@ -167,14 +167,14 @@ public class PayInfoRepositoryJDBCImpl implements PayInfoRepository<PayInfo,Job>
     }
 
     /**
-     * <strong>See >>></strong> {@link timeclock.interfaces.PayInfoRepository}.
+     * <strong>See >>></strong> {@link hu.gaborkolozsy.timeclock.daos.PayInfoRepository}.
      * 
      * @param project for selected the correct {@code Job}
      * 
      * @return {@code averageHourlyPay} data member of {@code PayInfo} 
      * object
      * @throws SQLException 
-     * @see timeclock.interfaces.PayInfoRepository
+     * @see hu.gaborkolozsy.timeclock.daos.PayInfoRepository
      * @deprecated <strong>not use this method</strong>
      */
     @Deprecated
@@ -189,14 +189,14 @@ public class PayInfoRepositoryJDBCImpl implements PayInfoRepository<PayInfo,Job>
     }
 
     /**
-     * <strong>See >>></strong> {@link timeclock.interfaces.PayInfoRepository}.
+     * <strong>See >>></strong> {@link hu.gaborkolozsy.timeclock.daos.PayInfoRepository}.
      * 
      * @param project for selected the correct {@code Job}
      * @param status for selected the correct {@code Job}
      * 
      * @return {@code totalPayment} data member of {@code PayInfo} object
      * @throws SQLException 
-     * @see timeclock.interfaces.PayInfoRepository
+     * @see hu.gaborkolozsy.timeclock.daos.PayInfoRepository
      * @deprecated <strong>not use this method</strong>
      */
     @Deprecated
@@ -212,13 +212,13 @@ public class PayInfoRepositoryJDBCImpl implements PayInfoRepository<PayInfo,Job>
     }
     
     /**
-     * <strong>See >>></strong> {@link timeclock.interfaces.PayInfoRepository}.
+     * <strong>See >>></strong> {@link hu.gaborkolozsy.timeclock.daos.PayInfoRepository}.
      * 
      * @param job for identification the correct {@code Job}
      * 
      * @return the {@code Job_id} for identification the correct {@code Job}
      * @throws SQLException 
-     * @see timeclock.interfaces.PayInfoRepository
+     * @see hu.gaborkolozsy.timeclock.daos.PayInfoRepository
      * @deprecated <strong>not use this method</strong>
      */
     @Deprecated
@@ -236,10 +236,10 @@ public class PayInfoRepositoryJDBCImpl implements PayInfoRepository<PayInfo,Job>
     }
     
     /**
-     * <strong>See >>></strong> {@link timeclock.interfaces.PayInfoRepository}.
+     * <strong>See >>></strong> {@link hu.gaborkolozsy.timeclock.daos.PayInfoRepository}.
      * 
      * @throws SQLException 
-     * @see timeclock.interfaces.PayInfoRepository
+     * @see hu.gaborkolozsy.timeclock.daos.PayInfoRepository
      */
     @Override
     public void close() throws SQLException {
