+++ TimeClock-MySQL/src/hu/gaborkolozsy/timeclock/model/Pay.java
@@ -2,7 +2,7 @@
  * Copyright (c) 2016, by Gábor Kolozsy. All rights reserved.
  * 
  */
-package timeclock.pay;
+package hu.gaborkolozsy.timeclock.model;
 
 /**
  * This object declared all datamembers for payment. Store in the database
